+++ tests/api/rest_patch_test.go
@@ -1,19 +1,60 @@
 package api
 
 import (
-	"fmt"
+	"encoding/json"
+	Invoice "github.com/starkinfra/core-go/tests/utils/invoice"
 	"github.com/starkinfra/core-go/starkcore/utils/hosts"
 	"github.com/starkinfra/core-go/starkcore/utils/rest"
 	"github.com/starkinfra/core-go/tests/utils"
+	"github.com/starkinfra/core-go/tests/utils/examples"
 	User "github.com/starkinfra/core-go/tests/utils/user"
 	"testing"
 )
 
 func TestSuccessPatch(t *testing.T) {
+	var invoice Invoice.Invoice
 	var patchData = map[string]interface{}{}
-	patchData["amount"] = 10
+	var amount = 200
+	patchData["amount"] = amount
 
-	invoice, err := rest.PatchId(
+	createdInvoice, err := Invoice.Create(examples.ExampleInvoice())
+	if err.Errors != nil {
+		t.Errorf("err: %s", err.Errors)
+	}
+
+	patch, err := rest.PatchId(
+		utils.SdkVersion,
+		hosts.Bank,
+		utils.ApiVersion,
+		utils.Language,
+		utils.Timeout,
+		User.ExampleProjectBank,
+		utils.ResourceInvoice,
+		createdInvoice[0].Id,
+		patchData,
+		nil,
+	)
+	if err.Errors != nil {
+		for _, e := range err.Errors {
+			t.Errorf("code: %s, message: %s", e.Code, e.Message)
+		}
+	}
+
+	unmarshalError := json.Unmarshal(patch, &invoice)
+	if unmarshalError != nil {
+		t.Errorf("unmarshalError: %s", unmarshalError)
+	}
+
+	if invoice.Amount != amount {
+		t.Errorf("invoice.Amount is not %d", amount)
+	}
+}
+
+func TestFailedPatch(t *testing.T) {
+	var patchData = map[string]interface{}{}
+	patchData["amount"] = 8
+
+	_, err := rest.PatchId(
 		utils.SdkVersion,
 		hosts.Bank,
 		utils.ApiVersion,
@@ -26,9 +67,7 @@ func TestSuccessPatch(t *testing.T) {
 		nil,
 	)
 	if err.Errors != nil {
-		for _, e := range err.Errors {
-			panic(fmt.Sprintf("code: %s, message: %s", e.Code, e.Message))
-		}
+		return
 	}
-	fmt.Println(string(invoice))
+	t.Errorf("unmarshalError: %s", err.Errors)
 }